@@ -5,7 +5,7 @@ homepage: https://www.agora.io
 repository: https://github.com/AgoraIO/Flutter-SDK
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.6"
 
 dependencies:
@@ -13,17 +13,17 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  json_annotation: ^3.0.0
+  json_annotation: ^4.0.0
 
 dev_dependencies:
+  integration_test: ^1.0.2+2
   flutter_test:
     sdk: flutter
-  e2e: ^0.2.1
   flutter_driver:
     sdk: flutter
-  pedantic: ^1.8.0
-  build_runner: ^1.7.0
-  json_serializable: ^3.2.0
+  pedantic: ^1.11.0
+  build_runner: ^1.11.5
+  json_serializable: ^4.0.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -41,15 +41,15 @@ flutter:
         pluginClass: AgoraRtcEnginePlugin
       ios:
         pluginClass: AgoraRtcEnginePlugin
-#      linux:
-#        pluginClass: AgoraRtcEnginePlugin
-#      macos:
-#        pluginClass: AgoraRtcEnginePlugin
-#      windows:
-#        pluginClass: AgoraRtcEnginePlugin
-#      web:
-#        fileName: agora_rtc_engine_web.dart
-#        pluginClass: AgoraRtcEngineWeb
+  #      linux:
+  #        pluginClass: AgoraRtcEnginePlugin
+  #      macos:
+  #        pluginClass: AgoraRtcEnginePlugin
+  #      windows:
+  #        pluginClass: AgoraRtcEnginePlugin
+  #      web:
+  #        fileName: agora_rtc_engine_web.dart
+  #        pluginClass: AgoraRtcEngineWeb
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
