@@ -10,20 +10,19 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:permission_handler/permission_handler.dart';
 
-const channelId0 = 'channel0';
-const channelId1 = 'channel1';
+const _channelId0 = 'channel0';
+const _channelId1 = 'channel1';
 
 /// MultiChannel Example
 class MultiChannel extends StatefulWidget {
-  RtcEngine _engine = null;
-  RtcChannel _channel0 = null, _channel1 = null;
-
   @override
   State<StatefulWidget> createState() => _State();
 }
 
 class _State extends State<MultiChannel> {
-  String renderChannelId;
+  late final RtcEngine _engine;
+  late final RtcChannel _channel0, _channel1;
+  String? renderChannelId;
   bool isJoined0 = false, isJoined1 = false;
   List<int> remoteUid0 = [], remoteUid1 = [];
 
@@ -36,17 +35,16 @@ class _State extends State<MultiChannel> {
   @override
   void dispose() {
     super.dispose();
-    widget._engine?.destroy();
+    _engine.destroy();
   }
 
   _initEngine() async {
-    widget._engine =
-        await RtcEngine.createWithConfig(RtcEngineConfig(config.appId));
+    _engine = await RtcEngine.createWithConfig(RtcEngineConfig(config.appId));
 
-    await widget._engine.enableVideo();
-    await widget._engine.startPreview();
-    await widget._engine.setChannelProfile(ChannelProfile.LiveBroadcasting);
-    await widget._engine.setClientRole(ClientRole.Broadcaster);
+    await _engine.enableVideo();
+    await _engine.startPreview();
+    await _engine.setChannelProfile(ChannelProfile.LiveBroadcasting);
+    await _engine.setClientRole(ClientRole.Broadcaster);
   }
 
   _joinChannel0() async {
@@ -54,12 +52,11 @@ class _State extends State<MultiChannel> {
       await [Permission.microphone, Permission.camera].request();
     }
 
-    widget._channel0 = await RtcChannel.create(channelId0);
-    this._addListener(widget._channel0);
+    _channel0 = await RtcChannel.create(_channelId0);
+    this._addListener(_channel0);
 
-    await widget._channel0.setClientRole(ClientRole.Broadcaster);
-    await widget._channel0
-        .joinChannel(null, null, 0, ChannelMediaOptions(true, true));
+    await _channel0.setClientRole(ClientRole.Broadcaster);
+    await _channel0.joinChannel(null, null, 0, ChannelMediaOptions(true, true));
   }
 
   _joinChannel1() async {
@@ -67,12 +64,11 @@ class _State extends State<MultiChannel> {
       await [Permission.microphone, Permission.camera].request();
     }
 
-    widget._channel1 = await RtcChannel.create(channelId1);
-    this._addListener(widget._channel1);
+    _channel1 = await RtcChannel.create(_channelId1);
+    this._addListener(_channel1);
 
-    await widget._channel1.setClientRole(ClientRole.Broadcaster);
-    await widget._channel1
-        .joinChannel(null, null, 0, ChannelMediaOptions(true, true));
+    await _channel1.setClientRole(ClientRole.Broadcaster);
+    await _channel1.joinChannel(null, null, 0, ChannelMediaOptions(true, true));
   }
 
   _addListener(RtcChannel channel) {
@@ -80,11 +76,11 @@ class _State extends State<MultiChannel> {
     channel.setEventHandler(
         RtcChannelEventHandler(joinChannelSuccess: (channel, uid, elapsed) {
       log('joinChannelSuccess ${channel} ${uid} ${elapsed}');
-      if (channelId == channelId0) {
+      if (channelId == _channelId0) {
         setState(() {
           isJoined0 = true;
         });
-      } else if (channelId == channelId1) {
+      } else if (channelId == _channelId1) {
         setState(() {
           isJoined1 = true;
         });
@@ -95,12 +91,12 @@ class _State extends State<MultiChannel> {
       log('userOffline ${channel.channelId} $uid $reason');
     }, leaveChannel: (stats) {
       log('leaveChannel ${channel.channelId} ${stats.toJson()}');
-      if (channelId == channelId0) {
+      if (channelId == _channelId0) {
         this.setState(() {
           isJoined0 = false;
           remoteUid0.clear();
         });
-      } else if (channelId == channelId1) {
+      } else if (channelId == _channelId1) {
         this.setState(() {
           isJoined1 = false;
           remoteUid1.clear();
@@ -109,21 +105,21 @@ class _State extends State<MultiChannel> {
     }, remoteVideoStateChanged: (uid, state, reason, elapsed) {
       log('remoteVideoStateChanged ${uid} ${state} ${reason} ${elapsed}');
       if (state == VideoRemoteState.Starting) {
-        if (channelId == channelId0) {
+        if (channelId == _channelId0) {
           this.setState(() {
             remoteUid0.add(uid);
           });
-        } else if (channelId == channelId1) {
+        } else if (channelId == _channelId1) {
           this.setState(() {
             remoteUid1.add(uid);
           });
         }
       } else if (state == VideoRemoteState.Stopped) {
-        if (channelId == channelId0) {
+        if (channelId == _channelId0) {
           this.setState(() {
             remoteUid0.removeWhere((element) => element == uid);
           });
-        } else if (channelId == channelId1) {
+        } else if (channelId == _channelId1) {
           this.setState(() {
             remoteUid1.removeWhere((element) => element == uid);
           });
@@ -133,21 +129,21 @@ class _State extends State<MultiChannel> {
   }
 
   _publishChannel0() async {
-    await widget._channel1?.unpublish();
-    await widget._channel0?.publish();
+    await _channel1.unpublish();
+    await _channel0.publish();
   }
 
   _publishChannel1() async {
-    await widget._channel0?.unpublish();
-    await widget._channel1?.publish();
+    await _channel0.unpublish();
+    await _channel1.publish();
   }
 
   _leaveChannel0() async {
-    await widget._channel0?.leaveChannel();
+    await _channel0.leaveChannel();
   }
 
   _leaveChannel1() async {
-    await widget._channel1?.leaveChannel();
+    await _channel1.leaveChannel();
   }
 
   @override
@@ -160,7 +156,7 @@ class _State extends State<MultiChannel> {
               children: [
                 Expanded(
                   flex: 1,
-                  child: RaisedButton(
+                  child: ElevatedButton(
                     onPressed: () {
                       if (isJoined0) {
                         this._leaveChannel0();
@@ -168,7 +164,7 @@ class _State extends State<MultiChannel> {
                         this._joinChannel0();
                       }
                     },
-                    child: Text('${isJoined0 ? 'Leave' : 'Join'} $channelId0'),
+                    child: Text('${isJoined0 ? 'Leave' : 'Join'} $_channelId0'),
                   ),
                 )
               ],
@@ -177,7 +173,7 @@ class _State extends State<MultiChannel> {
               children: [
                 Expanded(
                   flex: 1,
-                  child: RaisedButton(
+                  child: ElevatedButton(
                     onPressed: () {
                       if (isJoined1) {
                         this._leaveChannel1();
@@ -185,7 +181,7 @@ class _State extends State<MultiChannel> {
                         this._joinChannel1();
                       }
                     },
-                    child: Text('${isJoined1 ? 'Leave' : 'Join'} $channelId1'),
+                    child: Text('${isJoined1 ? 'Leave' : 'Join'} $_channelId1'),
                   ),
                 )
               ],
@@ -198,29 +194,29 @@ class _State extends State<MultiChannel> {
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: [
-              RaisedButton(
+              ElevatedButton(
                 onPressed: this._publishChannel0,
-                child: Text('Publish ${channelId0}'),
+                child: Text('Publish ${_channelId0}'),
               ),
-              RaisedButton(
+              ElevatedButton(
                 onPressed: () {
                   setState(() {
-                    renderChannelId = channelId0;
+                    renderChannelId = _channelId0;
                   });
                 },
-                child: Text('Render ${channelId0}'),
+                child: Text('Render ${_channelId0}'),
               ),
-              RaisedButton(
+              ElevatedButton(
                 onPressed: this._publishChannel1,
-                child: Text('Publish ${channelId1}'),
+                child: Text('Publish ${_channelId1}'),
               ),
-              RaisedButton(
+              ElevatedButton(
                 onPressed: () {
                   setState(() {
-                    renderChannelId = channelId1;
+                    renderChannelId = _channelId1;
                   });
                 },
-                child: Text('Render ${channelId1}'),
+                child: Text('Render ${_channelId1}'),
               ),
             ],
           ),
@@ -230,10 +226,10 @@ class _State extends State<MultiChannel> {
   }
 
   _renderVideo() {
-    List<int> remoteUid = null;
-    if (renderChannelId == channelId0) {
+    List<int>? remoteUid = null;
+    if (renderChannelId == _channelId0) {
       remoteUid = remoteUid0;
-    } else if (renderChannelId == channelId1) {
+    } else if (renderChannelId == _channelId1) {
       remoteUid = remoteUid1;
     }
     return Expanded(
