@@ -6,48 +6,48 @@ import 'package:flutter/widgets.dart';
 import './src/enums.dart';
 import './src/rtc_render_view.dart';
 
-/// (Android only) SurfaceView.
-///
-/// Please use [UIView](https://developer.apple.com/documentation/uikit/uiview) for iOS.
+/// SurfaceView.
 class SurfaceView extends RtcSurfaceView {
   /// Constructs a [SurfaceView]
   SurfaceView({
-    Key key,
-    bool zOrderMediaOverlay = false,
-    bool zOrderOnTop = false,
-    VideoRenderMode renderMode = VideoRenderMode.Hidden,
-    String channelId,
-    VideoMirrorMode mirrorMode = VideoMirrorMode.Auto,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
-    PlatformViewCreatedCallback onPlatformViewCreated,
+    Key? key,
+    String? channelId,
+    renderMode = VideoRenderMode.Hidden,
+    mirrorMode = VideoMirrorMode.Auto,
+    zOrderOnTop = false,
+    zOrderMediaOverlay = false,
+    PlatformViewCreatedCallback? onPlatformViewCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   }) : super(
-            key: key,
-            zOrderMediaOverlay: zOrderMediaOverlay,
-            zOrderOnTop: zOrderOnTop,
-            renderMode: renderMode,
-            channelId: channelId,
-            mirrorMode: mirrorMode,
-            gestureRecognizers: gestureRecognizers,
-            onPlatformViewCreated: onPlatformViewCreated,
-            uid: 0);
+          key: key,
+          uid: 0,
+          channelId: channelId,
+          renderMode: renderMode,
+          mirrorMode: mirrorMode,
+          zOrderOnTop: zOrderOnTop,
+          zOrderMediaOverlay: zOrderMediaOverlay,
+          onPlatformViewCreated: onPlatformViewCreated,
+          gestureRecognizers: gestureRecognizers,
+        );
 }
 
 /// (Android only) TextureView.
 class TextureView extends RtcTextureView {
   /// Constructs a [TextureView]
   TextureView({
-    Key key,
-    VideoRenderMode renderMode = VideoRenderMode.Hidden,
-    String channelId,
-    VideoMirrorMode mirrorMode = VideoMirrorMode.Auto,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
-    PlatformViewCreatedCallback onPlatformViewCreated,
+    Key? key,
+    String? channelId,
+    renderMode = VideoRenderMode.Hidden,
+    mirrorMode = VideoMirrorMode.Auto,
+    PlatformViewCreatedCallback? onPlatformViewCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   }) : super(
-            key: key,
-            renderMode: renderMode,
-            channelId: channelId,
-            mirrorMode: mirrorMode,
-            gestureRecognizers: gestureRecognizers,
-            onPlatformViewCreated: onPlatformViewCreated,
-            uid: 0);
+          key: key,
+          uid: 0,
+          channelId: channelId,
+          renderMode: renderMode,
+          mirrorMode: mirrorMode,
+          onPlatformViewCreated: onPlatformViewCreated,
+          gestureRecognizers: gestureRecognizers,
+        );
 }
