@@ -5,32 +5,30 @@ import 'package:flutter/services.dart';
 import './src/enums.dart';
 import './src/rtc_render_view.dart';
 
-/// (Android only) SurfaceView.
-///
-/// Please use [UIView](https://developer.apple.com/documentation/uikit/uiview) for iOS.
+/// SurfaceView.
 class SurfaceView extends RtcSurfaceView {
   /// Constructs a [SurfaceView]
   SurfaceView({
-    Key key,
-    @required int uid,
-    bool zOrderMediaOverlay = false,
-    bool zOrderOnTop = false,
-    VideoRenderMode renderMode = VideoRenderMode.Hidden,
-    String channelId,
-    VideoMirrorMode mirrorMode = VideoMirrorMode.Auto,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
-    PlatformViewCreatedCallback onPlatformViewCreated,
+    Key? key,
+    required int uid,
+    String? channelId,
+    renderMode = VideoRenderMode.Hidden,
+    mirrorMode = VideoMirrorMode.Auto,
+    zOrderOnTop = false,
+    zOrderMediaOverlay = false,
+    PlatformViewCreatedCallback? onPlatformViewCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   })  : assert(uid != 0),
         super(
           key: key,
           uid: uid,
-          zOrderMediaOverlay: zOrderMediaOverlay,
-          zOrderOnTop: zOrderOnTop,
-          renderMode: renderMode,
           channelId: channelId,
+          renderMode: renderMode,
           mirrorMode: mirrorMode,
-          gestureRecognizers: gestureRecognizers,
+          zOrderOnTop: zOrderOnTop,
+          zOrderMediaOverlay: zOrderMediaOverlay,
           onPlatformViewCreated: onPlatformViewCreated,
+          gestureRecognizers: gestureRecognizers,
         );
 }
 
@@ -38,21 +36,21 @@ class SurfaceView extends RtcSurfaceView {
 class TextureView extends RtcTextureView {
   /// Constructs a [TextureView]
   TextureView({
-    Key key,
-    @required int uid,
-    VideoRenderMode renderMode = VideoRenderMode.Hidden,
-    String channelId,
-    VideoMirrorMode mirrorMode = VideoMirrorMode.Auto,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
-    PlatformViewCreatedCallback onPlatformViewCreated,
+    Key? key,
+    required int uid,
+    String? channelId,
+    renderMode = VideoRenderMode.Hidden,
+    mirrorMode = VideoMirrorMode.Auto,
+    PlatformViewCreatedCallback? onPlatformViewCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   })  : assert(uid != 0),
         super(
           key: key,
           uid: uid,
-          renderMode: renderMode,
           channelId: channelId,
+          renderMode: renderMode,
           mirrorMode: mirrorMode,
-          gestureRecognizers: gestureRecognizers,
           onPlatformViewCreated: onPlatformViewCreated,
+          gestureRecognizers: gestureRecognizers,
         );
 }
