@@ -1,2 +1,2 @@
-export 'src/events.dart' show RtcChannelEventHandler;
+export 'src/events.dart' hide RtcEngineEventHandler;
 export 'src/rtc_channel.dart' show RtcChannel;
