@@ -1,5 +1,5 @@
 export 'src/classes.dart';
 export 'src/enums.dart';
-export 'src/events.dart' show RtcEngineEventHandler;
+export 'src/events.dart' hide RtcChannelEventHandler;
 export 'src/rtc_engine.dart' show RtcEngine;
 export 'src/rtc_engine_extension.dart';
