@@ -11,12 +11,10 @@ abstract class EnumConverter<E, T> {
 
   EnumConverter(this.e);
 
-  EnumConverter.fromValue(Map<E, T> map, T t) {
-    e = _$enumDecodeNullable(map, t);
-  }
+  EnumConverter.fromValue(Map<E, T> map, T t) : e = _$enumDecode(map, t);
 
   T toValue(Map<E, T> map) {
-    return map[e];
+    return map[e] as T;
   }
 }
 
