@@ -8,7 +8,7 @@ part of 'enum_converter.dart';
 
 AreaCodeConverter _$AreaCodeConverterFromJson(Map<String, dynamic> json) {
   return AreaCodeConverter(
-    _$enumDecodeNullable(_$AreaCodeEnumMap, json['e']),
+    _$enumDecode(_$AreaCodeEnumMap, json['e']),
   );
 }
 
@@ -17,36 +17,30 @@ Map<String, dynamic> _$AreaCodeConverterToJson(AreaCodeConverter instance) =>
       'e': _$AreaCodeEnumMap[instance.e],
     };
 
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
+K _$enumDecode<K, V>(
+  Map<K, V> enumValues,
+  Object? source, {
+  K? unknownValue,
 }) {
   if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
+    throw ArgumentError(
+      'A value must be provided. Supported values: '
+      '${enumValues.values.join(', ')}',
+    );
   }
 
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
-}
-
-T _$enumDecodeNullable<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
+  return enumValues.entries.singleWhere(
+    (e) => e.value == source,
+    orElse: () {
+      if (unknownValue == null) {
+        throw ArgumentError(
+          '`$source` is not one of the supported values: '
+          '${enumValues.values.join(', ')}',
+        );
+      }
+      return MapEntry(unknownValue, enumValues.values.first);
+    },
+  ).key;
 }
 
 const _$AreaCodeEnumMap = {
@@ -62,7 +56,7 @@ const _$AreaCodeEnumMap = {
 AudioCodecProfileTypeConverter _$AudioCodecProfileTypeConverterFromJson(
     Map<String, dynamic> json) {
   return AudioCodecProfileTypeConverter(
-    _$enumDecodeNullable(_$AudioCodecProfileTypeEnumMap, json['e']),
+    _$enumDecode(_$AudioCodecProfileTypeEnumMap, json['e']),
   );
 }
 
@@ -81,7 +75,7 @@ AudioEqualizationBandFrequencyConverter
     _$AudioEqualizationBandFrequencyConverterFromJson(
         Map<String, dynamic> json) {
   return AudioEqualizationBandFrequencyConverter(
-    _$enumDecodeNullable(_$AudioEqualizationBandFrequencyEnumMap, json['e']),
+    _$enumDecode(_$AudioEqualizationBandFrequencyEnumMap, json['e']),
   );
 }
 
@@ -107,7 +101,7 @@ const _$AudioEqualizationBandFrequencyEnumMap = {
 AudioLocalErrorConverter _$AudioLocalErrorConverterFromJson(
     Map<String, dynamic> json) {
   return AudioLocalErrorConverter(
-    _$enumDecodeNullable(_$AudioLocalErrorEnumMap, json['e']),
+    _$enumDecode(_$AudioLocalErrorEnumMap, json['e']),
   );
 }
 
@@ -129,7 +123,7 @@ const _$AudioLocalErrorEnumMap = {
 AudioLocalStateConverter _$AudioLocalStateConverterFromJson(
     Map<String, dynamic> json) {
   return AudioLocalStateConverter(
-    _$enumDecodeNullable(_$AudioLocalStateEnumMap, json['e']),
+    _$enumDecode(_$AudioLocalStateEnumMap, json['e']),
   );
 }
 
@@ -149,7 +143,7 @@ const _$AudioLocalStateEnumMap = {
 AudioMixingErrorCodeConverter _$AudioMixingErrorCodeConverterFromJson(
     Map<String, dynamic> json) {
   return AudioMixingErrorCodeConverter(
-    _$enumDecodeNullable(_$AudioMixingErrorCodeEnumMap, json['e']),
+    _$enumDecode(_$AudioMixingErrorCodeEnumMap, json['e']),
   );
 }
 
@@ -169,7 +163,7 @@ const _$AudioMixingErrorCodeEnumMap = {
 AudioMixingStateCodeConverter _$AudioMixingStateCodeConverterFromJson(
     Map<String, dynamic> json) {
   return AudioMixingStateCodeConverter(
-    _$enumDecodeNullable(_$AudioMixingStateCodeEnumMap, json['e']),
+    _$enumDecode(_$AudioMixingStateCodeEnumMap, json['e']),
   );
 }
 
@@ -189,7 +183,7 @@ const _$AudioMixingStateCodeEnumMap = {
 AudioOutputRoutingConverter _$AudioOutputRoutingConverterFromJson(
     Map<String, dynamic> json) {
   return AudioOutputRoutingConverter(
-    _$enumDecodeNullable(_$AudioOutputRoutingEnumMap, json['e']),
+    _$enumDecode(_$AudioOutputRoutingEnumMap, json['e']),
   );
 }
 
@@ -212,7 +206,7 @@ const _$AudioOutputRoutingEnumMap = {
 AudioProfileConverter _$AudioProfileConverterFromJson(
     Map<String, dynamic> json) {
   return AudioProfileConverter(
-    _$enumDecodeNullable(_$AudioProfileEnumMap, json['e']),
+    _$enumDecode(_$AudioProfileEnumMap, json['e']),
   );
 }
 
@@ -234,7 +228,7 @@ const _$AudioProfileEnumMap = {
 AudioRecordingQualityConverter _$AudioRecordingQualityConverterFromJson(
     Map<String, dynamic> json) {
   return AudioRecordingQualityConverter(
-    _$enumDecodeNullable(_$AudioRecordingQualityEnumMap, json['e']),
+    _$enumDecode(_$AudioRecordingQualityEnumMap, json['e']),
   );
 }
 
@@ -253,7 +247,7 @@ const _$AudioRecordingQualityEnumMap = {
 AudioRemoteStateConverter _$AudioRemoteStateConverterFromJson(
     Map<String, dynamic> json) {
   return AudioRemoteStateConverter(
-    _$enumDecodeNullable(_$AudioRemoteStateEnumMap, json['e']),
+    _$enumDecode(_$AudioRemoteStateEnumMap, json['e']),
   );
 }
 
@@ -274,7 +268,7 @@ const _$AudioRemoteStateEnumMap = {
 AudioRemoteStateReasonConverter _$AudioRemoteStateReasonConverterFromJson(
     Map<String, dynamic> json) {
   return AudioRemoteStateReasonConverter(
-    _$enumDecodeNullable(_$AudioRemoteStateReasonEnumMap, json['e']),
+    _$enumDecode(_$AudioRemoteStateReasonEnumMap, json['e']),
   );
 }
 
@@ -298,7 +292,7 @@ const _$AudioRemoteStateReasonEnumMap = {
 AudioReverbPresetConverter _$AudioReverbPresetConverterFromJson(
     Map<String, dynamic> json) {
   return AudioReverbPresetConverter(
-    _$enumDecodeNullable(_$AudioReverbPresetEnumMap, json['e']),
+    _$enumDecode(_$AudioReverbPresetEnumMap, json['e']),
   );
 }
 
@@ -331,7 +325,7 @@ const _$AudioReverbPresetEnumMap = {
 AudioReverbTypeConverter _$AudioReverbTypeConverterFromJson(
     Map<String, dynamic> json) {
   return AudioReverbTypeConverter(
-    _$enumDecodeNullable(_$AudioReverbTypeEnumMap, json['e']),
+    _$enumDecode(_$AudioReverbTypeEnumMap, json['e']),
   );
 }
 
@@ -352,7 +346,7 @@ const _$AudioReverbTypeEnumMap = {
 AudioSampleRateTypeConverter _$AudioSampleRateTypeConverterFromJson(
     Map<String, dynamic> json) {
   return AudioSampleRateTypeConverter(
-    _$enumDecodeNullable(_$AudioSampleRateTypeEnumMap, json['e']),
+    _$enumDecode(_$AudioSampleRateTypeEnumMap, json['e']),
   );
 }
 
@@ -371,7 +365,7 @@ const _$AudioSampleRateTypeEnumMap = {
 AudioScenarioConverter _$AudioScenarioConverterFromJson(
     Map<String, dynamic> json) {
   return AudioScenarioConverter(
-    _$enumDecodeNullable(_$AudioScenarioEnumMap, json['e']),
+    _$enumDecode(_$AudioScenarioEnumMap, json['e']),
   );
 }
 
@@ -395,7 +389,7 @@ const _$AudioScenarioEnumMap = {
 AudioVoiceChangerConverter _$AudioVoiceChangerConverterFromJson(
     Map<String, dynamic> json) {
   return AudioVoiceChangerConverter(
-    _$enumDecodeNullable(_$AudioVoiceChangerEnumMap, json['e']),
+    _$enumDecode(_$AudioVoiceChangerEnumMap, json['e']),
   );
 }
 
@@ -431,7 +425,7 @@ CameraCaptureOutputPreferenceConverter
     _$CameraCaptureOutputPreferenceConverterFromJson(
         Map<String, dynamic> json) {
   return CameraCaptureOutputPreferenceConverter(
-    _$enumDecodeNullable(_$CameraCaptureOutputPreferenceEnumMap, json['e']),
+    _$enumDecode(_$CameraCaptureOutputPreferenceEnumMap, json['e']),
   );
 }
 
@@ -451,7 +445,7 @@ const _$CameraCaptureOutputPreferenceEnumMap = {
 CameraDirectionConverter _$CameraDirectionConverterFromJson(
     Map<String, dynamic> json) {
   return CameraDirectionConverter(
-    _$enumDecodeNullable(_$CameraDirectionEnumMap, json['e']),
+    _$enumDecode(_$CameraDirectionEnumMap, json['e']),
   );
 }
 
@@ -469,7 +463,7 @@ const _$CameraDirectionEnumMap = {
 ChannelMediaRelayErrorConverter _$ChannelMediaRelayErrorConverterFromJson(
     Map<String, dynamic> json) {
   return ChannelMediaRelayErrorConverter(
-    _$enumDecodeNullable(_$ChannelMediaRelayErrorEnumMap, json['e']),
+    _$enumDecode(_$ChannelMediaRelayErrorEnumMap, json['e']),
   );
 }
 
@@ -497,7 +491,7 @@ const _$ChannelMediaRelayErrorEnumMap = {
 ChannelMediaRelayEventConverter _$ChannelMediaRelayEventConverterFromJson(
     Map<String, dynamic> json) {
   return ChannelMediaRelayEventConverter(
-    _$enumDecodeNullable(_$ChannelMediaRelayEventEnumMap, json['e']),
+    _$enumDecode(_$ChannelMediaRelayEventEnumMap, json['e']),
   );
 }
 
@@ -525,7 +519,7 @@ const _$ChannelMediaRelayEventEnumMap = {
 ChannelMediaRelayStateConverter _$ChannelMediaRelayStateConverterFromJson(
     Map<String, dynamic> json) {
   return ChannelMediaRelayStateConverter(
-    _$enumDecodeNullable(_$ChannelMediaRelayStateEnumMap, json['e']),
+    _$enumDecode(_$ChannelMediaRelayStateEnumMap, json['e']),
   );
 }
 
@@ -545,7 +539,7 @@ const _$ChannelMediaRelayStateEnumMap = {
 ChannelProfileConverter _$ChannelProfileConverterFromJson(
     Map<String, dynamic> json) {
   return ChannelProfileConverter(
-    _$enumDecodeNullable(_$ChannelProfileEnumMap, json['e']),
+    _$enumDecode(_$ChannelProfileEnumMap, json['e']),
   );
 }
 
@@ -563,7 +557,7 @@ const _$ChannelProfileEnumMap = {
 
 ClientRoleConverter _$ClientRoleConverterFromJson(Map<String, dynamic> json) {
   return ClientRoleConverter(
-    _$enumDecodeNullable(_$ClientRoleEnumMap, json['e']),
+    _$enumDecode(_$ClientRoleEnumMap, json['e']),
   );
 }
 
@@ -581,7 +575,7 @@ const _$ClientRoleEnumMap = {
 ConnectionChangedReasonConverter _$ConnectionChangedReasonConverterFromJson(
     Map<String, dynamic> json) {
   return ConnectionChangedReasonConverter(
-    _$enumDecodeNullable(_$ConnectionChangedReasonEnumMap, json['e']),
+    _$enumDecode(_$ConnectionChangedReasonEnumMap, json['e']),
   );
 }
 
@@ -613,7 +607,7 @@ const _$ConnectionChangedReasonEnumMap = {
 ConnectionStateTypeConverter _$ConnectionStateTypeConverterFromJson(
     Map<String, dynamic> json) {
   return ConnectionStateTypeConverter(
-    _$enumDecodeNullable(_$ConnectionStateTypeEnumMap, json['e']),
+    _$enumDecode(_$ConnectionStateTypeEnumMap, json['e']),
   );
 }
 
@@ -634,7 +628,7 @@ const _$ConnectionStateTypeEnumMap = {
 DegradationPreferenceConverter _$DegradationPreferenceConverterFromJson(
     Map<String, dynamic> json) {
   return DegradationPreferenceConverter(
-    _$enumDecodeNullable(_$DegradationPreferenceEnumMap, json['e']),
+    _$enumDecode(_$DegradationPreferenceEnumMap, json['e']),
   );
 }
 
@@ -653,7 +647,7 @@ const _$DegradationPreferenceEnumMap = {
 EncryptionModeConverter _$EncryptionModeConverterFromJson(
     Map<String, dynamic> json) {
   return EncryptionModeConverter(
-    _$enumDecodeNullable(_$EncryptionModeEnumMap, json['e']),
+    _$enumDecode(_$EncryptionModeEnumMap, json['e']),
   );
 }
 
@@ -673,7 +667,7 @@ const _$EncryptionModeEnumMap = {
 
 ErrorCodeConverter _$ErrorCodeConverterFromJson(Map<String, dynamic> json) {
   return ErrorCodeConverter(
-    _$enumDecodeNullable(_$ErrorCodeEnumMap, json['e']),
+    _$enumDecode(_$ErrorCodeEnumMap, json['e']),
   );
 }
 
@@ -766,7 +760,7 @@ const _$ErrorCodeEnumMap = {
 InjectStreamStatusConverter _$InjectStreamStatusConverterFromJson(
     Map<String, dynamic> json) {
   return InjectStreamStatusConverter(
-    _$enumDecodeNullable(_$InjectStreamStatusEnumMap, json['e']),
+    _$enumDecode(_$InjectStreamStatusEnumMap, json['e']),
   );
 }
 
@@ -793,7 +787,7 @@ const _$InjectStreamStatusEnumMap = {
 LastmileProbeResultStateConverter _$LastmileProbeResultStateConverterFromJson(
     Map<String, dynamic> json) {
   return LastmileProbeResultStateConverter(
-    _$enumDecodeNullable(_$LastmileProbeResultStateEnumMap, json['e']),
+    _$enumDecode(_$LastmileProbeResultStateEnumMap, json['e']),
   );
 }
 
@@ -812,7 +806,7 @@ const _$LastmileProbeResultStateEnumMap = {
 LighteningContrastLevelConverter _$LighteningContrastLevelConverterFromJson(
     Map<String, dynamic> json) {
   return LighteningContrastLevelConverter(
-    _$enumDecodeNullable(_$LighteningContrastLevelEnumMap, json['e']),
+    _$enumDecode(_$LighteningContrastLevelEnumMap, json['e']),
   );
 }
 
@@ -831,7 +825,7 @@ const _$LighteningContrastLevelEnumMap = {
 LocalVideoStreamErrorConverter _$LocalVideoStreamErrorConverterFromJson(
     Map<String, dynamic> json) {
   return LocalVideoStreamErrorConverter(
-    _$enumDecodeNullable(_$LocalVideoStreamErrorEnumMap, json['e']),
+    _$enumDecode(_$LocalVideoStreamErrorEnumMap, json['e']),
   );
 }
 
@@ -855,7 +849,7 @@ const _$LocalVideoStreamErrorEnumMap = {
 LocalVideoStreamStateConverter _$LocalVideoStreamStateConverterFromJson(
     Map<String, dynamic> json) {
   return LocalVideoStreamStateConverter(
-    _$enumDecodeNullable(_$LocalVideoStreamStateEnumMap, json['e']),
+    _$enumDecode(_$LocalVideoStreamStateEnumMap, json['e']),
   );
 }
 
@@ -874,7 +868,7 @@ const _$LocalVideoStreamStateEnumMap = {
 
 LogFilterConverter _$LogFilterConverterFromJson(Map<String, dynamic> json) {
   return LogFilterConverter(
-    _$enumDecodeNullable(_$LogFilterEnumMap, json['e']),
+    _$enumDecode(_$LogFilterEnumMap, json['e']),
   );
 }
 
@@ -895,7 +889,7 @@ const _$LogFilterEnumMap = {
 NetworkQualityConverter _$NetworkQualityConverterFromJson(
     Map<String, dynamic> json) {
   return NetworkQualityConverter(
-    _$enumDecodeNullable(_$NetworkQualityEnumMap, json['e']),
+    _$enumDecode(_$NetworkQualityEnumMap, json['e']),
   );
 }
 
@@ -919,7 +913,7 @@ const _$NetworkQualityEnumMap = {
 
 NetworkTypeConverter _$NetworkTypeConverterFromJson(Map<String, dynamic> json) {
   return NetworkTypeConverter(
-    _$enumDecodeNullable(_$NetworkTypeEnumMap, json['e']),
+    _$enumDecode(_$NetworkTypeEnumMap, json['e']),
   );
 }
 
@@ -942,7 +936,7 @@ const _$NetworkTypeEnumMap = {
 RtmpStreamingErrorCodeConverter _$RtmpStreamingErrorCodeConverterFromJson(
     Map<String, dynamic> json) {
   return RtmpStreamingErrorCodeConverter(
-    _$enumDecodeNullable(_$RtmpStreamingErrorCodeEnumMap, json['e']),
+    _$enumDecode(_$RtmpStreamingErrorCodeEnumMap, json['e']),
   );
 }
 
@@ -969,7 +963,7 @@ const _$RtmpStreamingErrorCodeEnumMap = {
 RtmpStreamingStateConverter _$RtmpStreamingStateConverterFromJson(
     Map<String, dynamic> json) {
   return RtmpStreamingStateConverter(
-    _$enumDecodeNullable(_$RtmpStreamingStateEnumMap, json['e']),
+    _$enumDecode(_$RtmpStreamingStateEnumMap, json['e']),
   );
 }
 
@@ -990,7 +984,7 @@ const _$RtmpStreamingStateEnumMap = {
 StreamFallbackOptionsConverter _$StreamFallbackOptionsConverterFromJson(
     Map<String, dynamic> json) {
   return StreamFallbackOptionsConverter(
-    _$enumDecodeNullable(_$StreamFallbackOptionsEnumMap, json['e']),
+    _$enumDecode(_$StreamFallbackOptionsEnumMap, json['e']),
   );
 }
 
@@ -1009,7 +1003,7 @@ const _$StreamFallbackOptionsEnumMap = {
 UserOfflineReasonConverter _$UserOfflineReasonConverterFromJson(
     Map<String, dynamic> json) {
   return UserOfflineReasonConverter(
-    _$enumDecodeNullable(_$UserOfflineReasonEnumMap, json['e']),
+    _$enumDecode(_$UserOfflineReasonEnumMap, json['e']),
   );
 }
 
@@ -1028,7 +1022,7 @@ const _$UserOfflineReasonEnumMap = {
 UserPriorityConverter _$UserPriorityConverterFromJson(
     Map<String, dynamic> json) {
   return UserPriorityConverter(
-    _$enumDecodeNullable(_$UserPriorityEnumMap, json['e']),
+    _$enumDecode(_$UserPriorityEnumMap, json['e']),
   );
 }
 
@@ -1046,7 +1040,7 @@ const _$UserPriorityEnumMap = {
 VideoCodecProfileTypeConverter _$VideoCodecProfileTypeConverterFromJson(
     Map<String, dynamic> json) {
   return VideoCodecProfileTypeConverter(
-    _$enumDecodeNullable(_$VideoCodecProfileTypeEnumMap, json['e']),
+    _$enumDecode(_$VideoCodecProfileTypeEnumMap, json['e']),
   );
 }
 
@@ -1065,7 +1059,7 @@ const _$VideoCodecProfileTypeEnumMap = {
 VideoFrameRateConverter _$VideoFrameRateConverterFromJson(
     Map<String, dynamic> json) {
   return VideoFrameRateConverter(
-    _$enumDecodeNullable(_$VideoFrameRateEnumMap, json['e']),
+    _$enumDecode(_$VideoFrameRateEnumMap, json['e']),
   );
 }
 
@@ -1088,7 +1082,7 @@ const _$VideoFrameRateEnumMap = {
 
 BitRateConverter _$BitRateConverterFromJson(Map<String, dynamic> json) {
   return BitRateConverter(
-    _$enumDecodeNullable(_$BitRateEnumMap, json['e']),
+    _$enumDecode(_$BitRateEnumMap, json['e']),
   );
 }
 
@@ -1105,7 +1099,7 @@ const _$BitRateEnumMap = {
 VideoMirrorModeConverter _$VideoMirrorModeConverterFromJson(
     Map<String, dynamic> json) {
   return VideoMirrorModeConverter(
-    _$enumDecodeNullable(_$VideoMirrorModeEnumMap, json['e']),
+    _$enumDecode(_$VideoMirrorModeEnumMap, json['e']),
   );
 }
 
@@ -1124,7 +1118,7 @@ const _$VideoMirrorModeEnumMap = {
 VideoOutputOrientationModeConverter
     _$VideoOutputOrientationModeConverterFromJson(Map<String, dynamic> json) {
   return VideoOutputOrientationModeConverter(
-    _$enumDecodeNullable(_$VideoOutputOrientationModeEnumMap, json['e']),
+    _$enumDecode(_$VideoOutputOrientationModeEnumMap, json['e']),
   );
 }
 
@@ -1143,7 +1137,7 @@ const _$VideoOutputOrientationModeEnumMap = {
 VideoQualityAdaptIndicationConverter
     _$VideoQualityAdaptIndicationConverterFromJson(Map<String, dynamic> json) {
   return VideoQualityAdaptIndicationConverter(
-    _$enumDecodeNullable(_$VideoQualityAdaptIndicationEnumMap, json['e']),
+    _$enumDecode(_$VideoQualityAdaptIndicationEnumMap, json['e']),
   );
 }
 
@@ -1162,7 +1156,7 @@ const _$VideoQualityAdaptIndicationEnumMap = {
 VideoRemoteStateConverter _$VideoRemoteStateConverterFromJson(
     Map<String, dynamic> json) {
   return VideoRemoteStateConverter(
-    _$enumDecodeNullable(_$VideoRemoteStateEnumMap, json['e']),
+    _$enumDecode(_$VideoRemoteStateEnumMap, json['e']),
   );
 }
 
@@ -1183,7 +1177,7 @@ const _$VideoRemoteStateEnumMap = {
 VideoRemoteStateReasonConverter _$VideoRemoteStateReasonConverterFromJson(
     Map<String, dynamic> json) {
   return VideoRemoteStateReasonConverter(
-    _$enumDecodeNullable(_$VideoRemoteStateReasonEnumMap, json['e']),
+    _$enumDecode(_$VideoRemoteStateReasonEnumMap, json['e']),
   );
 }
 
@@ -1209,7 +1203,7 @@ const _$VideoRemoteStateReasonEnumMap = {
 VideoRenderModeConverter _$VideoRenderModeConverterFromJson(
     Map<String, dynamic> json) {
   return VideoRenderModeConverter(
-    _$enumDecodeNullable(_$VideoRenderModeEnumMap, json['e']),
+    _$enumDecode(_$VideoRenderModeEnumMap, json['e']),
   );
 }
 
@@ -1229,7 +1223,7 @@ const _$VideoRenderModeEnumMap = {
 VideoStreamTypeConverter _$VideoStreamTypeConverterFromJson(
     Map<String, dynamic> json) {
   return VideoStreamTypeConverter(
-    _$enumDecodeNullable(_$VideoStreamTypeEnumMap, json['e']),
+    _$enumDecode(_$VideoStreamTypeEnumMap, json['e']),
   );
 }
 
@@ -1246,7 +1240,7 @@ const _$VideoStreamTypeEnumMap = {
 
 WarningCodeConverter _$WarningCodeConverterFromJson(Map<String, dynamic> json) {
   return WarningCodeConverter(
-    _$enumDecodeNullable(_$WarningCodeEnumMap, json['e']),
+    _$enumDecode(_$WarningCodeEnumMap, json['e']),
   );
 }
 
@@ -1294,7 +1288,7 @@ const _$WarningCodeEnumMap = {
 AudioChannelConverter _$AudioChannelConverterFromJson(
     Map<String, dynamic> json) {
   return AudioChannelConverter(
-    _$enumDecodeNullable(_$AudioChannelEnumMap, json['e']),
+    _$enumDecode(_$AudioChannelEnumMap, json['e']),
   );
 }
 
@@ -1316,7 +1310,7 @@ const _$AudioChannelEnumMap = {
 VideoCodecTypeConverter _$VideoCodecTypeConverterFromJson(
     Map<String, dynamic> json) {
   return VideoCodecTypeConverter(
-    _$enumDecodeNullable(_$VideoCodecTypeEnumMap, json['e']),
+    _$enumDecode(_$VideoCodecTypeEnumMap, json['e']),
   );
 }
 
@@ -1336,7 +1330,7 @@ const _$VideoCodecTypeEnumMap = {
 StreamPublishStateConverter _$StreamPublishStateConverterFromJson(
     Map<String, dynamic> json) {
   return StreamPublishStateConverter(
-    _$enumDecodeNullable(_$StreamPublishStateEnumMap, json['e']),
+    _$enumDecode(_$StreamPublishStateEnumMap, json['e']),
   );
 }
 
@@ -1356,7 +1350,7 @@ const _$StreamPublishStateEnumMap = {
 StreamSubscribeStateConverter _$StreamSubscribeStateConverterFromJson(
     Map<String, dynamic> json) {
   return StreamSubscribeStateConverter(
-    _$enumDecodeNullable(_$StreamSubscribeStateEnumMap, json['e']),
+    _$enumDecode(_$StreamSubscribeStateEnumMap, json['e']),
   );
 }
 
@@ -1376,7 +1370,7 @@ const _$StreamSubscribeStateEnumMap = {
 RtmpStreamingEventConverter _$RtmpStreamingEventConverterFromJson(
     Map<String, dynamic> json) {
   return RtmpStreamingEventConverter(
-    _$enumDecodeNullable(_$RtmpStreamingEventEnumMap, json['e']),
+    _$enumDecode(_$RtmpStreamingEventEnumMap, json['e']),
   );
 }
 
@@ -1394,7 +1388,7 @@ AudioSessionOperationRestrictionConverter
     _$AudioSessionOperationRestrictionConverterFromJson(
         Map<String, dynamic> json) {
   return AudioSessionOperationRestrictionConverter(
-    _$enumDecodeNullable(_$AudioSessionOperationRestrictionEnumMap, json['e']),
+    _$enumDecode(_$AudioSessionOperationRestrictionEnumMap, json['e']),
   );
 }
 
@@ -1415,7 +1409,7 @@ const _$AudioSessionOperationRestrictionEnumMap = {
 AudioEffectPresetConverter _$AudioEffectPresetConverterFromJson(
     Map<String, dynamic> json) {
   return AudioEffectPresetConverter(
-    _$enumDecodeNullable(_$AudioEffectPresetEnumMap, json['e']),
+    _$enumDecode(_$AudioEffectPresetEnumMap, json['e']),
   );
 }
 
@@ -1450,7 +1444,7 @@ const _$AudioEffectPresetEnumMap = {
 VoiceBeautifierPresetConverter _$VoiceBeautifierPresetConverterFromJson(
     Map<String, dynamic> json) {
   return VoiceBeautifierPresetConverter(
-    _$enumDecodeNullable(_$VoiceBeautifierPresetEnumMap, json['e']),
+    _$enumDecode(_$VoiceBeautifierPresetEnumMap, json['e']),
   );
 }
 
@@ -1479,7 +1473,7 @@ const _$VoiceBeautifierPresetEnumMap = {
 AudienceLatencyLevelTypeConverter _$AudienceLatencyLevelTypeConverterFromJson(
     Map<String, dynamic> json) {
   return AudienceLatencyLevelTypeConverter(
-    _$enumDecodeNullable(_$AudienceLatencyLevelTypeEnumMap, json['e']),
+    _$enumDecode(_$AudienceLatencyLevelTypeEnumMap, json['e']),
   );
 }
 
@@ -1496,7 +1490,7 @@ const _$AudienceLatencyLevelTypeEnumMap = {
 
 LogLevelConverter _$LogLevelConverterFromJson(Map<String, dynamic> json) {
   return LogLevelConverter(
-    _$enumDecodeNullable(_$LogLevelEnumMap, json['e']),
+    _$enumDecode(_$LogLevelEnumMap, json['e']),
   );
 }
 
@@ -1516,7 +1510,7 @@ const _$LogLevelEnumMap = {
 CaptureBrightnessLevelTypeConverter
     _$CaptureBrightnessLevelTypeConverterFromJson(Map<String, dynamic> json) {
   return CaptureBrightnessLevelTypeConverter(
-    _$enumDecodeNullable(_$CaptureBrightnessLevelTypeEnumMap, json['e']),
+    _$enumDecode(_$CaptureBrightnessLevelTypeEnumMap, json['e']),
   );
 }
 
@@ -1536,7 +1530,7 @@ const _$CaptureBrightnessLevelTypeEnumMap = {
 SuperResolutionStateReasonConverter
     _$SuperResolutionStateReasonConverterFromJson(Map<String, dynamic> json) {
   return SuperResolutionStateReasonConverter(
-    _$enumDecodeNullable(_$SuperResolutionStateReasonEnumMap, json['e']),
+    _$enumDecode(_$SuperResolutionStateReasonEnumMap, json['e']),
   );
 }
 
@@ -1556,7 +1550,7 @@ const _$SuperResolutionStateReasonEnumMap = {
 UploadErrorReasonConverter _$UploadErrorReasonConverterFromJson(
     Map<String, dynamic> json) {
   return UploadErrorReasonConverter(
-    _$enumDecodeNullable(_$UploadErrorReasonEnumMap, json['e']),
+    _$enumDecode(_$UploadErrorReasonEnumMap, json['e']),
   );
 }
 
@@ -1575,7 +1569,7 @@ const _$UploadErrorReasonEnumMap = {
 CloudProxyTypeConverter _$CloudProxyTypeConverterFromJson(
     Map<String, dynamic> json) {
   return CloudProxyTypeConverter(
-    _$enumDecodeNullable(_$CloudProxyTypeEnumMap, json['e']),
+    _$enumDecode(_$CloudProxyTypeEnumMap, json['e']),
   );
 }
 
@@ -1594,7 +1588,7 @@ const _$CloudProxyTypeEnumMap = {
 ExperienceQualityTypeConverter _$ExperienceQualityTypeConverterFromJson(
     Map<String, dynamic> json) {
   return ExperienceQualityTypeConverter(
-    _$enumDecodeNullable(_$ExperienceQualityTypeEnumMap, json['e']),
+    _$enumDecode(_$ExperienceQualityTypeEnumMap, json['e']),
   );
 }
 
@@ -1612,7 +1606,7 @@ const _$ExperienceQualityTypeEnumMap = {
 ExperiencePoorReasonConverter _$ExperiencePoorReasonConverterFromJson(
     Map<String, dynamic> json) {
   return ExperiencePoorReasonConverter(
-    _$enumDecodeNullable(_$ExperiencePoorReasonEnumMap, json['e']),
+    _$enumDecode(_$ExperiencePoorReasonEnumMap, json['e']),
   );
 }
 
